task/5 more work on validation
Introduced a new collection field method, hasFoos(), which allows an
empty collection from being disinguished from a missing colleciton. This
came up as necessary for validation of "enum" list in a Schema object,
which is required to be nonempty if it is present, but is not required
to be present.