task/5 Implemented checkUrl & checkEmail in ValBase
CheckUrl handles urls that may have variables, by replacing instances of
{...} with '1' and checking that the result is a valid URL. This won't
work if variables are allowed to cross syntactic boundaries in
URLs. Also, if this results in a scheme of '1:' it's replaced with a
special scheme for which we register a do-nothing protocol handler, so
URL class will allow it.