task/32 Fix harmless typo in sample program
The `validate` parameter on `processModel` was supposed to be passed to `OpenApiParser#parse` to suppress validation. The program worked as stated regardless, because of the later use of `validate` to control execution. That was done to prevent validation that would be triggered by calling `isValid()`, but it also meant the code produced the intended output. This way avoids the cost of validation, not just the evidence.