[#50] Partial fix
This PR prevents the NPE situation in the reported example, but it does
not properly handle `operationRef` values, and creates a validation
warning stating this.

The problem is that, in all of OpenAPI 3.0, this is the *one* place
where a reference string can appear outside a reference object! And
reference handling resolves ref strings in ref objects but does not have
a way of resolving other ref strings.

So this is a stop-gap until the necessary functionality can be added to
support this properly.