[#4] Start to implement serialization
The `toJson()` on `JsonOverlay` now checks whether its initial json is
still current, and if so just uses that. Thus for the vast majority of
applications that are read-only, serialization will be extremely
fast. However, the mutation API methods now all notice when an overlay
object's value has changed, and they mark their JSON value as no longer
current. Then `toJson()` calls `createJson()`, whose job it is to create
a new JSON structure representing the current overlay value. This is
implemented for all but `ObjectOverlay`, for which preserving ordering
of child structures is difficult and needs work.