[#4] Simple serializer works, but ignores refs
The `createJson` method now works for all overlay types and creates
correct JSON for each. However, it blindly follows references and so
does not yield the same result as a `toJson` method when the in the
presence of references, and when the parsed model has not been
modified. Following references should be an option, and without it,
`createJson` and `toJson` should yield equivalent trees.