Link API-Overview into other docs
Also switched back to using OAI/OpenAPI-Specification repo as source for
examples tests - and for the GettingStarted sample program, and switched
to using `OpenAPI3Parser` so the callback example won't fail from an
inability to determine OAS version (that example is (intentionally!)
missing `openapi` and `info` properties).

Made a few updates to the README and GettingStarted docs to reflect
current reality.