[#10] Use JavaParser for code generation
Rather than just assembling a bunch of strings, we now use JavaParser
AST classes to construct a compilation unit and then generate java
source from it. This sets the stage for the manual replacements called
for in this task.

We also move to Java8 as a minimum Java version for this component.