[#108] Improve support for subtyping
Generated factories are now subtype-aware. If type A extends type B,
then B's factory will determine the intended actual type to be created,
and, if that's not B, delegate to the intended type's factory.

Subtypes are determined by examining `extensionOf` properties in type
data.

When JSON is being parsed to create the object, the JSON structure must
include a "discriminator" that identifies the intended type. This is
managed via new `discriminator` and `discriminatorValue` properties in
type specifications. The `discriminator` should be a path from the JSON
from which the (supertype) is being instantiated (without initial
slash). The discriminator can be declared at top-level in type data, and
that will be the default for all types. The discriminator must provide a
string value from the JSON structure.

The `discriminatorValue` property in a (sub-)type is the value of the
discriminator obtained by the supertype factory, that will identify this
subtype as the intended type. Discriminator value defaults to the type's
name.