[#124] Implement new defautls
Also updated the types3.yaml file that specifies OAS3 types, to take
advantage of new defaults.

Type default isn't quite what was called for in the issue. Formerly, a
missing type used the field name, if that named another type in the type
configuration. (E.g. the field named "Info" would use the "Info" type as
its type). So now that rule is applied first, but if the field name does
not name a configured type, "String" is used instead.