[#126] Redid interface generator
Doing it the way i should have from the beginning... no more pointless
Member interfaces with separate paramters for parts of the member
declaration. JavaParser can easily parse them whole, and with xtend it's
easy to create them whole. And it results in far better readability!

Now on to impl generator