[#128] Added simple enum support
A type configured with an `enumValues` property (a list of enum constant
names) is generated as follows:

* In the `models` directory, an enum class named for the type and with
  the given member constants
* In the `impl` directory, an impl class that extends `EnumOverlay<E>`,
  where `E` is the generated enum type.

All other properties of such a type in the type config are ignored. It's
probably possible to add manual members to the enum class that will be
preserved during regen, but adding constructor args is not currently
supported, so there's probably not much of use that can be gained.