[#131] Reference cycles now handled properly
Two main changes accommodate cycles.

First, when the factory method that creates an overlay from a JsonNode
now has an overridable method, `createAndRegister`, whose default is the
former non-reference case of the create method. The
`PropertiesOverlayFactory` class overrides `createAndRegister`, with the
following strategy:

1. The factory first creates an object from an empty JsonNode (usually
   an ObjectNode, but it's an ArrayNode for SecurityParameters.
2. That "placeholder" overlay object is registered in the reference
   registry as the overlay corresponding to that empty JsonNode.
3. Next, the factory creates an object from the actual JsonNode passed
   to the factory method.
4. Finally, the contents of this second object are copied into the
   placeholder overlay.

The key here is that step 1 does not cause any child objects to be
created, because JSON container is empty. By the time references to this
object are encountered during elaboration in step 3, the reference has a
registered overlay, which is returned by the factory without attempting
to create another instance. All the linkages are thereby satisfied, but
the object is empty and not properly wired up behind the scenes until
step 4.

Second, the validator needs to guard against cycles in its traversal of
the model. A thread local IdentityHashMap is used to keep track of
overlays that have been (or are currently being) validated, and any
attempt to validate an object already in that map returns without doing
anything.

SecurityParameters presented an odd challenge, since its JsonNode is an
array, and it's the same array as is used to instantiate its sole
child (its "parameters" list). Step 2 was assocating the array node with
a SecurityParametersImpl object, but then in step 3, the ListOverlay
creation when elaborating the `parameters` child was stumbling over
that, because it would try to use the registered overlay and caused a
ClassCastException when trying to cast the SecurityParametersImpl to a
ListOverlay. The solution was to make OverlayFactory use the reference
registry value only if that value is compatible with the factory's
declared class (using the factory's already available `isCompatible`
method)