[#180] First error integrated with position info
This demonstrates the API by which positional info will be presented by
the validator, i.e. via ValidationItem#getPositionInfo()

Only a single validation currently adheres to this interface, namely a
check that the `type` property in a Schema is among the allowed
types. Working with a model that includes, e.g., a schema with `type:
xxx` should threfore produce a validation item with a non-null and
correct PositionInfo value.

PositionInfo really needs to include information about the source URL,
but it currently does not. That will be fixed along the way.