ZEN-4322: Update OAS3 parser to support OAS 3.0.1
From
https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#serverVariableObject
:
> enum	[string]	An enumeration of string values to be used if the
substitution options are from a limited set.

> default	string	REQUIRED. The default value to use for substitution,
and to send, if an alternate value is not supplied. Unlike the Schema
Object's default, this value MUST be provided by the consumer.

From
https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#linkObject
:
> requestBody	Any | {expression}	A literal value or {expression} to use
as a request body when calling the target operation.

* no "headers" - request headers should be declared in "parameters"

From
https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.1.md#schemaObject

No "examples" (plural), only "example"(singular):
> example	Any	A free-form property to include an example of an instance
for this schema. To represent examples that cannot be naturally
represented in JSON or YAML, a string value can be used to contain the
example with escaping where necessary.