[#186] Update model definition
Also regened models and updated affected validation classes.

This and subsequent work on this issue are being done while consulting
PR #182, branch task/ZEN-4322b. Unfortuantely, that PR was opened just
prior to a major overhaul of the validation framework, that was being
done simultaneously. As a result, the PR diffs are nearly useless.

What I'm doing instead is using github to view a diff of current that PR
branch agaisnt commit 6e2cd8e7d29319dc1f10e35a7f63154fb0e44243, which is
the commit just prior to the renovation of validation. I'll use that
much more manageable diff to guide the changes in this branch.