[#26] First cut at a validation catalog
YAML file produced from a careful review of the 3.0.2 specification,
capturing level-2 validations in model object & field specifications.
Level-3 validations appear in free text blocks attached to object
defintions. Those not appearing in the specification but seemingly
needed, are marked with "inSpecification: false". There are also
"notes", mostly attached to object definitions, describing issues that
maybe ought to be addressed in future revisions of the specification.