[#209] Allow relative URLs where permitted
When an initial attempt to create a new URL from a URL field fails, the
code now checks whether relative URLs are allowed, and if so, tries
again - this time supplying a context URL against which the given URL is
resolved. If this fails, a specific error message is provided indicating
that a relative URL is not permitted here. When relative URLs *are*
allowed, a generic "bad url" message, incorporating the URL error
message, is provided.

This required a change to the URL field validation method signatures, so
a handful of validators were affected.